+++ src/modules/dashboard/Dashboard.js
@@ -57,33 +57,41 @@ const Dashboard = ({
         <Grid.Column mobile={8} computer={4}>
           <StatsCard
             icon="users"
-            color="blue"
+            customColor="#2D9CDB"
             text="Customers Today"
-            amount={widgetData.totalNumberOfCustomers}
+            amount={widgetData.totalNumberOfCustomers
+              .toString()
+              .replace(/\B(?=(\d{3})+(?!\d))/g, ',')}
           />
         </Grid.Column>
         <Grid.Column mobile={8} computer={4}>
           <StatsCard
             icon="boxes"
-            color="orange"
+            customColor="#F2994A"
             text="Products Sold Today"
-            amount={widgetData.totalNumberOfProducts}
+            amount={widgetData.totalNumberOfProducts
+              .toString()
+              .replace(/\B(?=(\d{3})+(?!\d))/g, ',')}
           />
         </Grid.Column>
         <Grid.Column mobile={8} computer={4}>
           <StatsCard
             icon="credit card"
-            color="teal"
+            customColor="#219653"
             text="Revenue Today"
-            amount={`€ ${widgetData.totalGrossIncome}`}
+            amount={`${widgetData.totalGrossIncome
+              .toString()
+              .replace(/\B(?=(\d{3})+(?!\d))/g, ',')} €`}
           />
         </Grid.Column>
         <Grid.Column mobile={8} computer={4}>
           <StatsCard
             icon="tag"
-            color="pink"
+            customColor="#9B51E0"
             text="Monthly Revenue"
-            amount={`€ ${widgetData.totalMonthlyGrossIncome}`}
+            amount={`${widgetData.totalMonthlyGrossIncome
+              .toString()
+              .replace(/\B(?=(\d{3})+(?!\d))/g, ',')} €`}
           />
         </Grid.Column>
       </Grid.Row>