+++ src/modules/dashboard/components/AlertsToolbar.js
@@ -45,37 +45,35 @@ const Toolbar = ({
     changeDate(date);
   };
   return (
-    <Segment className="toolbar">
-      <Grid stackable>
-        <Grid.Row verticalAlign="middle" columns="equal">
-          <Grid.Column mobile={16} computer={4}>
-            <DatePicker
-              type="range"
-              onChange={handleDateChange}
-              className="full-width"
-            />
-          </Grid.Column>
-          <Grid.Column mobile={16} computer={3}>
-            <Dropdown
-              placeholder="Kiosks"
-              selection
-              className="full-width"
-              onChange={handleKioskChange}
-              options={kiosks}
-            />
-          </Grid.Column>
-          <Grid.Column mobile={16} computer={4}>
-            <Dropdown
-              placeholder="All Alerts"
-              selection
-              options={alertsOptions}
-              className="full-width"
-              onChange={handleAlertsChange}
-            />
-          </Grid.Column>
-        </Grid.Row>
-      </Grid>
-    </Segment>
+    <Grid stackable>
+      <Grid.Row verticalAlign="middle" columns="equal">
+        <Grid.Column mobile={16} computer={4}>
+          <DatePicker
+            type="range"
+            onChange={handleDateChange}
+            className="full-width"
+          />
+        </Grid.Column>
+        <Grid.Column mobile={16} computer={4}>
+          <Dropdown
+            placeholder="All Kiosks"
+            selection
+            className="full-width"
+            onChange={handleKioskChange}
+            options={kiosks}
+          />
+        </Grid.Column>
+        <Grid.Column mobile={16} computer={4}>
+          <Dropdown
+            placeholder="All Alerts"
+            selection
+            options={alertsOptions}
+            className="full-width"
+            onChange={handleAlertsChange}
+          />
+        </Grid.Column>
+      </Grid.Row>
+    </Grid>
   );
 };