+++ src/modules/kiosks/components/ActivityLog.js
@@ -30,8 +30,13 @@ const columns = [
     field: 'created',
     formatter: ({ created }) => {
       let [date, time] = created.split(' ');
-      return `${date}, ${time}`;
+      return (
+        <div style={{ textAlign: 'left', width: '120px' }}>
+          {date}, {time}
+        </div>
+      );
     },
+    className: 'ActivityLog',
   },
   {
     title: 'Event',
@@ -45,9 +50,9 @@ const columns = [
         event.touchedScales.length > 0
       ) {
         return (
-          `Products Touched -` +
+          `Products Taken -` +
           event.touchedScales.map(scl => {
-            return ` Weight: ${scl.weight}g / Cable Id: ${scl.id}`;
+            return ` Weight: ${scl.weight}g / Cable ID: ${scl.id}`;
           })
         );
       } else if (
@@ -55,7 +60,7 @@ const columns = [
         event.touchedScales !== undefined &&
         event.touchedScales.length === 0
       )
-        return `Products Touched - Empty`;
+        return `Products Taken - Empty`;
       else if (
         event.paymentTerminal !== null &&
         event.paymentTerminal !== undefined
@@ -140,7 +145,7 @@ const ActivityLogGrid = ({
               <Header.Content>Activity Log</Header.Content>
             </Header>
           </SegmentHeader>
-          <Grid>
+          <Grid stackable>
             <Grid.Row className="activity-log-filter-row">
               <Grid.Column mobile={16} computer={4}>
                 <DatePicker type="range" onChange={handleDateChange} />