+++ src/modules/products/actions/index.js
@@ -10,6 +10,8 @@ export const deleteProductSaga = createAction('@@saga/DELETE_PRODUCT');
 export const modifyProductImage = createAction('@@saga/MODIFY_PRODUCT_IMAGE');
 export const deleteProductImage = createAction('@@saga/DELETE_PRODUCT_IMAGE');
 export const getPriceHistory = createAction('@@saga/GET_PRICE_HISTORY');
+export const resetPriceHistory = createAction('@@saga/RESET_PRICE_HISTORY');
+export const deleteActivePriceHistory = createAction('@@saga/DELETE_PRICE_HISTORY');
 
 // State Actions
 export const getProductListSuccess = createAction('@@state/GET_PRODUCTS_LIST_SUCCESS');
@@ -23,3 +25,5 @@ export const deleteProductSuccess = createAction('@@state/DELETE_PRODUCT_SUCCESS
 export const modifyProductImageSuccess = createAction('@@saga/MODIFY_PRODUCT_IMAGE_SUCCESS');
 export const deleteProductImageSuccess = createAction('@@saga/DELETE_PRODUCT_IMAGE_SUCCESS');
 export const getPriceHistorySuccess = createAction('@@saga/GET_PRICE_HISTORY_SUCCESS');
+export const resetPriceHistorySuccess = createAction('@@saga/RESET_PRICE_HISTORY_SUCCESS');
+export const deleteActivePriceHistorySuccess = createAction('@@saga/DELETE_PRICE_HISTORY_SUCCESS');