+++ src/modules/transactions/ProductList.js
@@ -95,9 +95,7 @@ const ProductList = ({
     getAllProducts({ data });
     getProductsWidgetsData({ ...widgetPayload });
   };
-  useEffect(() => {
-    getProductsWidgetsData();
-  }, []);
+
   useEffect(() => {
     getData({ sort });
   }, [page, perPage, kiosk, dateRange, product]);
@@ -113,40 +111,48 @@ const ProductList = ({
       />
       <Grid>
         <Grid.Row stretched className="custom-widgets">
-          <Grid.Column mobile={8} computer={4}>
+          <Grid.Column mobile={16} computer={4} tablet={8}>
             <StatsCard
               icon="star"
-              color="green"
+              customColor="#219653"
               text="Most Sold Product"
               amount={`${widgetsData.mostRefilledProductName} `}
               secondaryText={`${widgetsData.mostRefilledProductValue.sum}  Sold`}
+              popup
+              fontTo18
             />
           </Grid.Column>
-          <Grid.Column mobile={8} computer={4}>
+          <Grid.Column mobile={16} computer={4} tablet={8}>
             <StatsCard
               icon="long arrow alternate down"
-              color="orange"
-              text="Least Sold product"
+              customColor="#F2994A"
+              text="Least Sold Product"
               amount={`${widgetsData.leastSoldProductName} `}
               secondaryText={`${widgetsData.leastSoldProductValue.sum}  Sold`}
+              popup
+              fontTo18
             />
           </Grid.Column>
-          <Grid.Column mobile={8} computer={4}>
+          <Grid.Column mobile={16} computer={4} tablet={8}>
             <StatsCard
               icon="dolly flatbed"
-              color="blue"
+              customColor="#2D9CDB"
               text="Most Refilled Product"
               amount={`${widgetsData.mostRefilledProductName} `}
               secondaryText={`${widgetsData.mostRefilledProductValue.sum}  Refilled`}
+              popup
+              fontTo18
             />
           </Grid.Column>
-          <Grid.Column mobile={8} computer={4}>
+          <Grid.Column mobile={16} computer={4} tablet={8}>
             <StatsCard
               icon="reply"
-              color="violet"
-              text="Most Removed Products"
+              customColor="#9B51E0"
+              text="Most Removed Product"
               amount={`${widgetsData.mostRemovedProductName} `}
               secondaryText={`${widgetsData.mostRemovedProductValue.sum}  Removed`}
+              popup
+              fontTo18
             />
           </Grid.Column>
         </Grid.Row>
@@ -158,6 +164,7 @@ const ProductList = ({
         setSortByInCaller={sort => setSort([sort])}
         // sortDirection="DESC"
       />
+      <br></br>
       <Pagination
         totalCount={total}
         page={page}