+++ src/modules/transactions/components/ProductsToolbar.js
@@ -3,18 +3,10 @@ import { func, arrayOf, object } from 'prop-types';
 import { Grid, Dropdown } from 'semantic-ui-react';
 import { exportCsvProducts } from '../actions';
 import { connect } from 'react-redux';
-import CustomButton from 'modules/shared/components/CustomButton';
 import { toast } from 'react-semantic-toasts';
 import DatePicker from 'modules/shared/components/Datepicker';
 
-const Toolbar = ({
-  kiosks,
-  // productsListValue,
-  changeDate,
-  changeKiosk,
-  changeProduct,
-  exportCsvProducts,
-}) => {
+const Toolbar = ({ kiosks, changeDate, changeKiosk, exportCsvProducts }) => {
   const [exportData, changeExportData] = useState(false);
 
   const handleDateChange = value => {
@@ -52,68 +44,26 @@ const Toolbar = ({
       kiosk: value,
     });
   };
-  const handleProductChange = (e, { value }) => {
-    changeProduct(value);
-  };
-
-  const DownloadCsv = () => {
-    if (exportData.from == '' && exportData.to == '') {
-      window.alert('Bitte wählen Sie zuerst das Datum.');
-    } else {
-      let value = {
-        from: Math.round(new Date(exportData.from)),
-        to: Math.round(new Date(exportData.to)),
-        kioskId: exportData.kioskId ? exportData.kioskId : '',
-      };
-      exportCsvProducts(value);
-      toast({
-        description: 'Downloading the requested file.',
-        animation: 'fade left',
-        icon: 'info',
-        color: 'blue',
-      });
-    }
-  };
 
   return (
     <div
       style={{
-        margin: '20px 0',
+        marginTop: '20px',
       }}
     >
       <Grid stackable>
         <Grid.Row verticalAlign="middle">
-          <Grid.Column mobile={16} computer={3}>
+          <Grid.Column mobile={16} tablet={8} computer={3}>
             <DatePicker type="range" onChange={handleDateChange} />
           </Grid.Column>
-          <Grid.Column mobile={16} computer={3}>
+          <Grid.Column mobile={16} tablet={8} computer={3}>
             <Dropdown
-              placeholder="Kiosk"
+              placeholder="All Kiosks"
               selection
               options={kiosks}
               className="full-width"
               onChange={handleKioskChange}
             />
-          </Grid.Column>
-          <Grid.Column mobile={16} computer={3}>
-            {/* <Dropdown
-              placeholder="All Products"
-              selection
-              options={productsListValue}
-              className="full-width"
-              onChange={handleProductChange}
-            /> */}
-          </Grid.Column>
-
-          <Grid.Column computer={3}></Grid.Column>
-          <Grid.Column mobile={16} computer={3}>
-            {/* <CustomButton
-              label="Download CSV&nbsp;"
-              icon="arrow down icon"
-              className="custom-button-default"
-              onClick={DownloadCsv}
-              disabled={!Boolean(exportData)}
-            /> */}
           </Grid.Column>
         </Grid.Row>
       </Grid>